Changed the syntax of import from IMPORT id FROM id; to IMPORT id FROM "quoted/text/lib"; Writing external METHOD calls via the EXTERNAL methname(arglist); now works.
Against my better judgement in a few cases,
cleaned up the old EXT statement code from Abbott
to work in a typesafe manner. A better solution
is to get the CALL syntax working as defined in 
the grammar, but this takes a good deal more work
and meanwhile people want to get their phds...

Still some known bugs in relation_util.
I need to commit the test models that go with
the EXT statements. Fixed a bunch of lintish
compiler complaints. Don't try to solve external
models yet.


git-svn-id: https://svn.ascend4.org/trunk@467 f15cd148-d149-4907-926e-73cf92b3ee03