Renamed ideal-solution test files to follow more consistent naming convention
In the future, `init_mix' will be the test file used, and remembering history will be handled with revisions rather than by coping over to new files.  Of course, it is sometimes useful to have several versions simultaneously, to test them side-by-side.

Transferred functions used in `init_mix' to a dedicated file, and function prototypes to dedicated header file, to reduce clutter.


git-svn-id: https://svn.ascend4.org/branches/jacob@2947 f15cd148-d149-4907-926e-73cf92b3ee03